python_campaign_launcher/requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Task queue
celery==5.3.4
redis==5.0.1

# AI and content generation
google-generativeai==0.3.2
openai==1.3.9

# Video and image processing
moviepy==1.0.3
Pillow==10.1.0
opencv-python==4.8.1.78

# Audio processing
pydub==0.25.1
gTTS==2.4.0

# Social media APIs
requests==2.31.0
httpx==0.25.2

# File storage
boto3==1.34.0
minio==7.2.0

# Configuration and data
PyYAML==6.0.1
python-dotenv==1.0.0

# Database (optional)
sqlalchemy==2.0.23
alembic==1.13.0

# Utilities
aiofiles==23.2.1
python-multipart==0.0.6
jinja2==3.1.2
croniter==2.0.1

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0